+++ js/data.js
@@ -1,0 +1,57 @@
+const MAX_PRICE = 200;
+const MAX_ROOMS_COUNT = 20;
+const MAX_GUESTS_COUNT = 20;
+
+const ACCOMMODATION_TYPES = [
+  'palace',
+  'flat',
+  'house',
+  'bungalow',
+  'hotel'
+];
+const CHECK_IN_TIME = [
+  '12:00',
+  '13:00',
+  '14:00'
+];
+const CHECK_OUT_TIME = [
+  '12:00',
+  '13:00',
+  '14:00'
+];
+const POSSIBLE_FEATURES = [
+  'wifi',
+  'dishwasher',
+  'parking',
+  'washer',
+  'elevator',
+  'conditioner'
+];
+const POSSIBLE_PHOTOS = [
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/duonguyen-8LrGtIxxa4w.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/brandon-hoogenboom-SNxQGWxZQi0.jpg',
+  'https://assets.htmlacademy.ru/content/intensive/javascript-1/keksobooking/claire-rendall-b6kAwr1i0Iw.jpg'
+];
+const POSSIBLE_TITLES = [
+  'Такого вы еще не видели',
+  'Попробуйте что-то новое',
+  'Хорошее предложение'
+];
+const POSSIBLE_DESCRIPTIONS =[
+  'Предложение говорит само за себя',
+  'Станция метро в 5 минутах пешком',
+  'Из окна открываются отличные виды, большой потенциал для активного отдыха'
+];
+
+export {
+  MAX_PRICE,
+  MAX_ROOMS_COUNT,
+  MAX_GUESTS_COUNT,
+  ACCOMMODATION_TYPES,
+  CHECK_IN_TIME,
+  CHECK_OUT_TIME,
+  POSSIBLE_FEATURES,
+  POSSIBLE_PHOTOS,
+  POSSIBLE_TITLES,
+  POSSIBLE_DESCRIPTIONS
+};
